requirements.txt
aiohappyeyeballs==2.4.2
aiohttp==3.10.6
aiosignal==1.3.1
argcomplete==3.2.3
attrs==24.2.0
blinker==1.7.0
click==8.1.7
colorama==0.4.6
discord==2.3.2
discord.py==2.4.0
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
frozenlist==1.4.1
greenlet==3.0.3
idna==3.10
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
multidict==6.1.0
packaging==24.0
pipx==1.5.0
platformdirs==4.2.0
python-dotenv==1.0.1
SQLAlchemy==2.0.29
typing_extensions==4.10.0
userpath==1.9.2
Werkzeug==3.0.2
yarl==1.13.0
